File organization for StripeModel.
